+++ tests/test_weo/test_scraper.py
@@ -2,7 +2,6 @@
 
 import pytest
 from unittest.mock import patch, Mock
-import requests
 from bs4 import BeautifulSoup
 import io
 from zipfile import ZipFile, BadZipFile
@@ -12,28 +11,6 @@ from imf_reader.config import NoDataError
 
 
 TEST_URL = "https://test.com"
-
-
-def test_make_request():
-    """Test make_request"""
-
-    # test successful request
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.status_code = 200
-        response = scraper.make_request(TEST_URL)
-        assert response == mock_get.return_value
-
-    # test failed request
-    with patch("requests.get") as mock_get:
-        mock_get.side_effect = requests.exceptions.RequestException
-        with pytest.raises(ConnectionError, match="Could not connect to"):
-            scraper.make_request(TEST_URL)
-
-    # test when status code is not 200
-    with patch("requests.get") as mock_get:
-        mock_get.return_value.status_code = 404
-        with pytest.raises(ConnectionError, match="Could not connect to"):
-            scraper.make_request(TEST_URL)
 
 
 def test_get_soup():
